trying to compile

+++ core/src/main/scala/org/apache/spark/rdd/RDDCheckpointData.scala
@@ -121,7 +121,7 @@ private[spark] class RDDCheckpointData[T: ClassTag](@transient rdd: RDD[T])
       logInfo("Finished checkpointing RDD " + rdd.id + " to " + path + ", new parent is RDD " + newRDD.id)
     }
 
-    return (Int) ((end_time - start_time)/1000)
+    return ((end_time - start_time)/1000)
   }
 
   // Do the checkpointing of the RDD. Called after the first job using that RDD is over.
